Tema3/Verde/R0004/pyproject.toml
[project]
name = "reinas"
version = "0.1.0"
description = ""
authors = [
    {name = "Verde"}
]
requires-python = ">=3.12"
dependencies = [
    "pyyaml (>=6.0.3,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
